server/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
pyasn1==0.6.0
pycparser==2.21
requests==2.31.0
rsa==4.9
urllib3==2.0.7
python-dotenv==1.0.1
pycryptodome==3.20.0
